refactor: 전역상태 관리 파일 atom으로 통일
