+++ src/pages/ProductDetail/atom.ts
